requirements.txt
flask==2.3.3
python-dotenv
twilio
google-generativeai
requests
mcp
google-genai